Remove the unroll type class
I wasn't terribly satisfied with it. It's certainly possible to restore
int he future, but it doesn't make a ton of sense either in the context
of stacked monads since you have a single return argument anyway.

Nevertheless, this is a little bit of a limitation of the Cont2
hierarchy that it's difficult to write a type for the printer. As you
need to provide this `Nothing` argument as the failure continuation. It
means in turn you need to know the base of the input stack.